// ui/src/types/bindings/notification.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Notification = { "type": "StartedFeedRefresh", feed_uuid: string, } | { "type": "FinishedFeedRefresh", feed_uuid: string, } | { "type": "EntriesFetched", feed_uuid: string, entry_uuids: Array<string>, };